backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
mistralai>=1.12.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.6
sse-starlette==2.1.3
httpx==0.28.1
